move TokenId into token.rs, initial Token trait
