+++ consensus/enclave/mock/src/mock_consensus_enclave.rs
@@ -14,8 +14,7 @@ use mc_consensus_enclave_api::{
 use mc_crypto_keys::{Ed25519Public, X25519Public};
 use mc_sgx_report_cache_api::{ReportableEnclave, Result as SgxReportResult};
 use mc_transaction_core::{
-    tx::{TokenId, TxOutMembershipProof},
-    Block, BlockContents, BlockSignature,
+    tx::TxOutMembershipProof, Block, BlockContents, BlockSignature, TokenId,
 };
 use std::collections::BTreeMap;
