+++ consensus/enclave/src/lib.rs
@@ -22,8 +22,7 @@ use mc_sgx_report_cache_api::{ReportableEnclave, Result as ReportableEnclaveResu
 use mc_sgx_types::{sgx_enclave_id_t, sgx_status_t, *};
 use mc_sgx_urts::SgxEnclave;
 use mc_transaction_core::{
-    tx::{TokenId, TxOutMembershipProof},
-    Block, BlockContents, BlockSignature,
+    tx::TxOutMembershipProof, Block, BlockContents, BlockSignature, TokenId,
 };
 use std::{collections::BTreeMap, path, result::Result as StdResult, sync::Arc};
