+++ consensus/service/src/config.rs
@@ -7,7 +7,7 @@ use mc_common::{HashMap, HashSet, NodeID, ResponderId};
 use mc_consensus_enclave::FeeMap;
 use mc_consensus_scp::{QuorumSet, QuorumSetMember};
 use mc_crypto_keys::{DistinguishedEncoding, Ed25519Pair, Ed25519Private};
-use mc_transaction_core::tx::TokenId;
+use mc_transaction_core::TokenId;
 use mc_util_parse::parse_duration_in_seconds;
 use mc_util_uri::{
     AdminUri, ConnectionUri, ConsensusClientUri as ClientUri, ConsensusPeerUri as PeerUri,
