+++ transaction/core/src/lib.rs
@@ -19,6 +19,7 @@ mod amount;
 mod blockchain;
 mod domain_separators;
 mod memo;
+mod token;
 mod tx_error;
 
 pub mod constants;
@@ -37,6 +38,7 @@ pub mod proptest_fixtures;
 pub use amount::{get_value_mask, Amount, AmountError, Commitment, CompressedCommitment};
 pub use blockchain::*;
 pub use memo::{EncryptedMemo, MemoError, MemoPayload};
+pub use token::{tokens, Token, TokenId};
 pub use tx::MemoContext;
 pub use tx_error::{NewMemoError, NewTxError};
